Player Freezing Effect (#10)
* Optimised Heat Source Block Calculation

Executing location.clone().add() is very slow.
Doing this for radius cubed times per game tick is adding up to a
massive delay.
Reworked this to do only simple integer calculation and comparisons.

* Added Taking Damage Once Temperature Critical

When temperature falls below the critical threshold, player also
receives damage.

* Resetting Temperature Once Player Dies

* Fixed Temperature Overlay And Damage Effect

* Added Session Storage For Actual Player Ticks

Technically, the player freeze ticks are not actual ticks anymore.
The real actual ticks are used to calculate physical and visual effect.
Those are now stored and loaded whenever a player joins or quits the
game.

* Fixed Bug Where Player Was Not Freezing

The player was not freezing after reconnecting, when a heat source was
destroyed. Fixed this issue by storing the clamped value, instead of a
"free" one.

* Code Refactoring

Refactored some code to use new language features.
Might not be in the sense of data encapsulation.
Renamed TemperatureManager to TemperatureHandler for more consistency.

* Exposing Interpolation Functions In Config

The temperature interpolation functions are now exposed in the config
file, so that users can determine how fast a player freezes or warms up.