Exposing Leather Armour Isolation Values In Config
All values are in the range of 0 to 100, expressing per cent values.
It is now configurable how much leather armour can isolate in total,
default 80%.
Each individual armour part is also configurable, as how much the
contribute to the total of 80% reduction.
It is recommended to back up the current configuration, let the plugin
create a new one and copy non-default values over.