More Isolation Armour Item Support
As many items as the server host wants to give an isolation value to.
The item, however, must be a wearable item, since only items that are
worn will contribute to the players isolation value.
Item naming convention in the config is to be taken from the PaperMC
"Material" class (JavaDoc).
Values have to be configured as integers in the range from 0 to 100.