Players Clothing Isolates (#11)
* Added Basic Leather Armour Isolation

Changed architecture of temperature system to use double to allow for
reduction rates introduced by leather armour.

* Exposing Leather Armour Isolation Values In Config

All values are in the range of 0 to 100, expressing per cent values.
It is now configurable how much leather armour can isolate in total,
default 80%.
Each individual armour part is also configurable, as how much the
contribute to the total of 80% reduction.
It is recommended to back up the current configuration, let the plugin
create a new one and copy non-default values over.

* More Isolation Armour Item Support

As many items as the server host wants to give an isolation value to.
The item, however, must be a wearable item, since only items that are
worn will contribute to the players isolation value.
Item naming convention in the config is to be taken from the PaperMC
"Material" class (JavaDoc).
Values have to be configured as integers in the range from 0 to 100.

* Refactored Isolation Code & Added Insurance

Ensuring that the item defined in the isolation configuration is an as
armour wearable item.

Refactored the code to have less indentation.